Decorated.Logger (Decorated v0.1.0)

A simple decorator that takes a message and logs the function name, arguments, and result of a function call.

Example

defmodule MyMath do
  use Decorated.Logger

  @decorate info_log(catch: :error)
  @spec add(number(), number()) :: number()
  def add(a, b), do: {:ok, a + b}

  @decorate debug_log(message: fn result -> "sub() -> %{result}" end)
  @spec sub(number(), number()) :: number()
  def sub(a, b), do: {:ok, a - b}

  @decorate warning_log(message: fn result -> "multiply() -> %{result}" end)
  @spec multiply(number(), number()) :: number()
  def multiply(a, b), do: {:ok, a * b}

  @decorate silent_log(catch: :error)
  @spec divide!(number(), number()) :: number()
  def divide!(_, 0), do: raise("do not divide by zero in MyMath functions")
  def divide!(a, b), do: a / b
end

iex> MyMath.add(1, 2)
#> 23:14:44.410 [info] MyMath.add!(1, 2) ->
{:ok, 3}

iex> MyMath.sub(1, 2)
#> 23:14:44.410 [debug] sub() -> -1
{:ok, -1}

iex> MyMath.multiply(1, 2)
#> 23:14:44.410 [warning] multiply() -> 2
{:ok, 2}

iex> MyMath.divide!(4, 2)
2.0

iex> MyMath.divide!(4, 0)
** (RuntimeError) do not divide by zero in MyMath functions

#> 23:14:44.410 [error] MyMath.divide!(4, 0) -> ** (RuntimeError) do not divide by zero in MyMath functions

Types

options()

@type options() :: [Decorated.Logger.Opts.log_opt()]

Functions

alert_log()
(macro)

@spec alert_log() :: no_return()

A decorator that produces a ALERT level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

alert_log(var1)
(macro)

@spec alert_log(options()) :: no_return()

A decorator that produces a ALERT level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

critical_log()
(macro)

@spec critical_log() :: no_return()

A decorator that produces a CRITICAL level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

critical_log(var1)
(macro)

@spec critical_log(options()) :: no_return()

A decorator that produces a CRITICAL level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

debug_log()
(macro)

@spec debug_log() :: no_return()

A decorator that produces a DEBUG level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

debug_log(var1)
(macro)

@spec debug_log(options()) :: no_return()

A decorator that produces a DEBUG level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

emergency_log()
(macro)

@spec emergency_log() :: no_return()

A decorator that produces a EMERGENCY level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

emergency_log(var1)
(macro)

@spec emergency_log(options()) :: no_return()

A decorator that produces a EMERGENCY level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

error_log()
(macro)

@spec error_log() :: no_return()

A decorator that produces a ERROR level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

error_log(var1)
(macro)

@spec error_log(options()) :: no_return()

A decorator that produces a ERROR level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

info_log()
(macro)

@spec info_log() :: no_return()

A decorator that produces an INFO level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

info_log(var1)
(macro)

@spec info_log(options()) :: no_return()

A decorator that produces an INFO level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

notice_log()
(macro)

@spec notice_log() :: no_return()

A decorator that produces a NOTICE level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

notice_log(var1)
(macro)

@spec notice_log(options()) :: no_return()

A decorator that produces a NOTICE level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

silent_log()
(macro)

@spec silent_log() :: no_return()

A decorator that produces no log UNLESS a configuration is provided in which case it produces a log at the provided level using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

silent_log(var1)
(macro)

@spec silent_log(options()) :: no_return()

A decorator that produces no log UNLESS a configuration is provided in which case it produces a log at the provided level using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

warning_log()
(macro)

@spec warning_log() :: no_return()

A decorator that produces a WARNING level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.

warning_log(var1)
(macro)

@spec warning_log(options()) :: no_return()

A decorator that produces a WARNING level log using the provided message.

The message has interpolated access to the function arguments by name defined in the imeadiately following function definition.