Decorated.Logger.Opts (Decorated v0.1.0)

A module that defines the configurations options for the Decorated.Logger decorator.

Option

:message - message_opt() - The message to log. This can be a string, a binary, or an IO list.

:none - log_level() - The log level to use when the decorated function returns :none. Defaults to the same level as the decorating log level.

:error - log_level() - The log level to use when the decorated function returns :error or {:error, _}. Defaults to the same level as the decorating log level.

:catch - log_level() - The log level to use when the decorated function raises an Exception.kind(). Defaults to the same level as the decorating log level.

:ignored - ignore_behaviour() - The behaviour to use when the decorated function has an ignored argument ex. _ or _ignored. Defaults to :rename.

:drop - Drop the ignored argument from the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(:ok)

:rename - Rename the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(_, _a, :ok)

:keep - Keep the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(nil, nil, :ok)

:metadata - Keyword.t() - The metadata to include in the log message. Injects the :line metadata by default.

Example

defmodule MyMath do
  use Decorated.Logger

  @decorate info_log(catch: :error)
  @spec add(number(), number()) :: number()
  def add(a, b), do: {:ok, a + b}

  @decorate debug_log(error: :critical)
  @spec sub(number(), number()) :: number()
  def sub(a, b), do: {:ok, a - b}

  @decorate critical_log(message: fn result -> "multiply() -> %{result}" end)
  @spec multiply(number(), number()) :: number()
  def multiply(a, b), do: {:ok, a * b}
end

iex> MyMath.add(1, 2)
#> 23:14:44.410 [info] MyMath.add!(1, 2) -> {:ok, 3}
{:ok, 3}

iex> MyMath.sub(1, 2)
#> 23:14:44.410 [debug] sub() -> -1
{:ok, -1}

iex> MyMath.multiply(1, 2)
#> 23:14:44.410 [critical] multiply() -> 2
{:ok, 2}

Types

ignore_behaviour()

@type ignore_behaviour() :: :drop | :rename | :keep

The ignore behaviour to use for a given configuration.

:drop - Drop the ignored argument from the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(:ok)

:rename - Rename the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(_, _a, :ok)

:keep - Keep the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(nil, nil, :ok)

level()

@type level() :: :silent | Logger.level()

The log level to use for a given configuration.

In addition to the standard log levels provided by Logger, the following log levels are also supported:

:silent - Do not log the decorated function.

log_opt()

@type log_opt() ::
  message_opt()
  | {:none, level()}
  | {:error, level()}
  | {:catch, level()}
  | {:ignored, ignore_behaviour()}
  | {:metadata, Keyword.t()}

The log level configuration, message, and metadata, and log level overides for common divergent scenarios.

Option

:message - message_opt() - The message to log. This can be a string, a binary, or an IO list.

:none - log_level() - The log level to use when the decorated function returns :none. Defaults to the same level as the decorating log level.

:error - log_level() - The log level to use when the decorated function returns :error or {:error, _}. Defaults to the same level as the decorating log level.

:catch - log_level() - The log level to use when the decorated function raises an Exception.kind(). Defaults to the same level as the decorating log level.

:ignored - ignore_behaviour() - The behaviour to use when the decorated function has an ignored argument ex. _ or _ignored. Defaults to :rename.

:drop - Drop the ignored argument from the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(:ok)

:rename - Rename the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(_, _a, :ok)

:keep - Keep the ignored argument in the log message. foo(_, _a, bar)(nil, nil, :ok) => foo(nil, nil, :ok)

:metadata - Keyword.t() - The metadata to include in the log message. Injects the :line metadata by default.

message_opt()

@type message_opt() ::
  {:message,
   String.t()
   | (any() -> String.t())
   | (any(), Decorator.Decorate.Context.t() -> String.t())}

Functions

build!(level, opts, ctx)

Builds a Decorated.Logger.Opts struct from the given options.